Added Random Player
The Random Player is added as a new object in IPlayer class. Selecting Player 2 as the player would enable this player since we have not set up the multiplayer functionality.
The Player runs on a while loop for now and should be changed so it only makes one move per instant, possibly with an if statement of some sort.